Add healthchekc test
