+++ deputy-claimer/test/integration/kava/initstate/.kava/config/write-file-atomic-3890029616445916362
@@ -1,0 +1,4 @@
+{
+	"key": "60276907d26f1538963be718",
+	"addrs": []
+}
